Rename: constants 패키지를 domain 하위 패키지로 변경
